Fix debugger error formatting in circuit panel (#3191)
This PR fixes https://github.com/microsoft/qdk/issues/1536

Previously, runtime errors from the debugger were displayed as
unreadable text in the circuit panel. The errors were being serialized
as plain JS `Error` objects in the WASM layer and the extension was
displaying the raw stack trace instead of the actual error message. This
PR aligns the debugger's error handling with the circuit panel command
by serializing errors as structured `IQSharpError` diagnostics all the
way from WASM through the worker boundary to the VS Code extension,
where they are now formatted with the same `errorsToHtml` function used
by the "Show Circuit" command — including error codes, source locations,
and clickable links to the relevant Q# source files.

## Before:
<img width="1262" height="499" alt="image"
src="https://github.com/user-attachments/assets/5dbddb77-9d46-46ab-be3c-b12e314e71f3"
/>


## Now:
<img width="732" height="485" alt="image"
src="https://github.com/user-attachments/assets/27464c87-6daf-4ed2-9711-16377aabf1e0"
/>
