+++ source/vscode/src/debugger/session.ts
@@ -27,12 +27,18 @@ import {
   IStructStepResult,
   IVariable,
   IVariableChild,
+  QdkDiagnostics,
   QscEventTarget,
   StepResultId,
   log,
 } from "qsharp-lang";
 import { updateCircuitPanel } from "../circuit";
-import { basename, isQdkDocument, toVsCodeRange } from "../common";
+import {
+  basename,
+  isQdkDocument,
+  toVsCodeRange,
+  errorsToHtml,
+} from "../common";
 import {
   DebugEvent,
   EventType,
@@ -1109,16 +1115,20 @@ export class QscDebugSession extends LoggingDebugSession {
       this.config.showCircuit ||
       isPanelOpen("circuit", this.program.projectName)
     ) {
-      // Error returned from the debugger has a message and a stack (which also includes the message).
-      // We would ideally retrieve the original runtime error, and format it to be consistent
-      // with the other runtime errors that can be shown in the circuit panel, but that will require
-      // a bit of refactoring.
-      const stack =
-        error && typeof error === "object" && typeof error.stack === "string"
-          ? escapeHtml(error.stack)
-          : undefined;
+      let errorHtml: string | undefined;
+      if (error instanceof QdkDiagnostics) {
+        errorHtml = errorsToHtml(error.diagnostics);
+      } else if (error) {
+        const message = error instanceof Error ? error.message : String(error);
+        errorHtml = `<pre>${escapeHtml(message)}</pre>`;
+      }
 
-      const circuit = await this.debugService.getCircuit();
+      let circuit;
+      try {
+        circuit = await this.debugService.getCircuit();
+      } catch (e) {
+        log.error("Failed to retrieve circuit from debug service: %O", e);
+      }
 
       updateCircuitPanel(
         this.program.profile,
@@ -1126,7 +1136,7 @@ export class QscDebugSession extends LoggingDebugSession {
         !this.revealedCircuit,
         {
           circuit,
-          errorHtml: stack ? `<pre>${stack}</pre>` : undefined,
+          errorHtml,
           simulated: true,
         },
       );
